+++ packages/kysely/src/users/update.ts
@@ -1,14 +1,7 @@
 import { PostUsersBody } from "@authhero/adapter-interfaces";
 import { Kysely } from "kysely";
 import { Database } from "../db";
-
-function getEmailVerified(user: Partial<PostUsersBody>): number | undefined {
-  if (user.email_verified === undefined) {
-    return undefined;
-  }
-
-  return user.email_verified ? 1 : 0;
-}
+import { flattenObject } from "../utils/flatten";
 
 export function update(db: Kysely<Database>) {
   return async (
@@ -16,20 +9,16 @@ export function update(db: Kysely<Database>) {
     user_id: string,
     user: Partial<PostUsersBody>,
   ): Promise<boolean> => {
-    const sqlUser = {
+    const sqlUser = flattenObject({
       ...user,
-      email_verified: getEmailVerified(user),
       updated_at: new Date().toISOString(),
-    };
-
-    // TODO: this currently overwrites the entire app_metadata and user_metadata. Should it be merged instead?
-    if (user.app_metadata) {
-      sqlUser.app_metadata = JSON.stringify(user.app_metadata);
-    }
-
-    if (user.user_metadata) {
-      sqlUser.user_metadata = JSON.stringify(user.user_metadata);
-    }
+      app_metadata: user.app_metadata
+        ? JSON.stringify(user.app_metadata)
+        : undefined,
+      user_metadata: user.user_metadata
+        ? JSON.stringify(user.user_metadata)
+        : undefined,
+    });
 
     const results = await db
       .updateTable("users")